Divided AssemblyEntityBase into OperationAssemblyEntity and IOAssemblyEntity Refactored cast methods in assembly Generic changes in assembly Moved Clone Debug Example XML file to separate directory in XMLFiles Fixed issue of field order in transform component Added implementation of subphase for handling phases of subjob component Added null check in prepare method for operation property Removed RuntimeException from prepare method Removed componentAssemblyMap.properties Fixed compile time error in UnionAllAssembly Added hydrograph.execution.engine properties for tez and mr in constants class. Refactored code of PlatformHelper : corrected checking of hydrograph.execution.engine value. Added changed for changing the output file of debug from avro to text Fixed issue of applying data type on header when quotes are present in data

+++ hydrograph.engine/hydrograph.engine.cascading/src/main/java/hydrograph/engine/cascading/assembly/generator/RunProgramComponentGenerator.java
@@ -35,7 +35,7 @@ public class RunProgramComponentGenerator extends CommandComponentGeneratorBase 
 	@Override
 	public void initializeEntity() {
 		runProgramEntity.setComponentId(runProgram.getId());
-		runProgramEntity.setPhase(runProgram.getPhase().intValue());
+		runProgramEntity.setPhase(runProgram.getPhase());
 		runProgramEntity.setCommand(runProgram.getCommand().getValue());
 	}
